+++ ui-core/src/types/forms.ts
@@ -142,6 +142,7 @@ export type I18nSurveyElement = string | I18nMap
 export type FormElement = FormPanel | HtmlElement | Question
 
 export type FormPanel = BaseElement & {
+  title: string
   type: 'panel'
   elements: (HtmlElement | Question)[]
 }
